# Mammoth_Labs/blockscan-api/internal/adapter/http/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Backend Template",
        "title": "blockscan-go API",
        "contact": {},
        "license": {
            "name": "Go Echo"
        },
        "version": "1.0"
    },
    "host": "tracking.gmmtchain.io",
    "basePath": "/api/v1",
    "paths": {
        "/chain/token/{chain_id}/{contract_address}": {
            "get": {
                "description": "GetToken / token 관련 정보를 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chain"
                ],
                "summary": "Get Token / *공용",
                "operationId": "GetToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "컨트랙트 주소",
                        "name": "contract_address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_chain_Chain"
                        }
                    }
                }
            }
        },
        "/chain/tokens/{chain_id}": {
            "get": {
                "description": "GetTokens / token 관련 정보들을 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chain"
                ],
                "summary": "Get Tokens / *공용",
                "operationId": "GetTokens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_chain_Chain"
                        }
                    }
                }
            }
        },
        "/chain/{chain_id}": {
            "get": {
                "description": "GetChain / chain 관련 정보를 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chain"
                ],
                "summary": "Get Chain / *공용",
                "operationId": "GetChain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_chain_Chain"
                        }
                    }
                }
            }
        },
        "/coin/price": {
            "get": {
                "description": "Get / 코인 가격 받아오는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crypto Currency"
                ],
                "summary": "Get coins price",
                "operationId": "GetPrice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_crypto_currency_GetCryptoCurrencyData"
                        }
                    }
                }
            }
        },
        "/contract/wallet/nft/{wallet_address}": {
            "get": {
                "description": "ContractGetWalletNFTsByCollection / 월렛이 보유한 NFT들을 컬렉션 별로 분류해서 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Get wallet nfts by collection / *공용",
                "operationId": "ContractGetWalletNFTsByCollection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "지갑주소",
                        "name": "wallet_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "가져올 아이템 수 최대 100",
                        "name": "take",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "건너뛸 아이템 수 (옵션) 기본 0",
                        "name": "skip",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_contract_GetContractWalletNFTsByCollectionData"
                        }
                    }
                }
            }
        },
        "/contract/{hash}/wallet/nft/{wallet_address}": {
            "get": {
                "description": "ContractGetCollectionNFTsForWallet / 해당 컨트랙트에 월렛이 보유한 NFT들을 반환하는 API / *공용",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Contract"
                ],
                "summary": "Get collection nfts for wallet / *공용",
                "operationId": "ContractGetCollectionNFTsForWallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x api key",
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "컨트랙트 주소",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "지갑주소",
                        "name": "wallet_address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "블록체인 체인넘버",
                        "name": "chain_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "가져올 아이템 수 최대 100",
                        "name": "take",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "건너뛸 아이템 수 (옵션) 기본 0",
                        "name": "skip",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_contract_GetContractCollectionNFTsForWalletData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_chain_Chain": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_chain.Chain"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-array_blockscan-go_internal_core_domain_contract_GetContractWalletNFTsByCollectionData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_contract.GetContractWalletNFTsByCollectionData"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
...
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_chain_Chain": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/blockscan-go_internal_core_domain_chain.Chain"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_contract_GetContractCollectionNFTsForWalletData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/blockscan-go_internal_core_domain_contract.GetContractCollectionNFTsForWalletData"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
        "blockscan-go_internal_adapter_http_response.ApiResponseSwagger-blockscan-go_internal_core_domain_crypto_currency_GetCryptoCurrencyData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/blockscan-go_internal_core_domain_crypto_currency.GetCryptoCurrencyData"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/blockscan-go_internal_core_common_utils.ErrorType"
                }
            }
        },
...
        "blockscan-go_internal_core_common_utils.ErrorType": {
            "type": "string",
            "enum": [
                "No data found",
                "Syntax",
                "Duplicate",
                "Invalid sql statement",
                "Bad connection",
                "Invalid character",
                "Foreign key",
                "Check violation",
                "Success",
                "Fail",
                "Not found",
                "Invalid input",
                "Database",
                "Internal"
            ],
            "x-enum-varnames": [
                "NoDataFound",
                "SyntaxError",
                "DuplicateError",
                "InvalidSQLStatementError",
                "BadConnectionError",
                "JsonParseError",
                "ForeignKeyError",
                "CheckViolationError",
                "Success",
                "Fail",
                "NotFound",
                "InvalidInput",
                "DatabaseError",
                "InternalError"
            ]
        },
        "blockscan-go_internal_core_domain_chain.Chain": {
            "type": "object",
            "properties": {
                "chain_id": {
                    "type": "integer"
                },
                "contract_address": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "scan_site": {
                    "type": "string"
                },
                "site": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_domain_contract.ErcList": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "aws_image_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "token_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_domain_contract.GetContractCollectionNFTsForWalletData": {
            "type": "object",
            "properties": {
                "aws_banner_image": {
                    "type": "string"
                },
                "aws_logo_image": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "erc_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_contract.ErcList"
                    }
                },
                "erc_type": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "wallet_address": {
                    "type": "string"
                },
                "wallet_nickname": {
                    "type": "string"
                },
                "wallet_profile": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_domain_contract.GetContractWalletNFTsByCollectionData": {
            "type": "object",
            "properties": {
                "aws_banner_image": {
                    "type": "string"
                },
                "aws_logo_image": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "erc_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/blockscan-go_internal_core_domain_contract.ErcList"
                    }
                },
                "erc_type": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "tokens_amount": {
                    "type": "string"
                },
                "unique_tokens_count": {
                    "type": "string"
                },
                "wallet_address": {
                    "type": "string"
                },
                "wallet_nickname": {
                    "type": "string"
                },
                "wallet_profile": {
                    "type": "string"
                }
            }
        },
        "blockscan-go_internal_core_domain_crypto_currency.GetCryptoCurrencyData": {
            "type": "object",
            "properties": {
                "bnb_price": {
                    "type": "string"
                },
                "eth_price": {
                    "type": "string"
                },
                "gmmt_price": {
                    "type": "string"
                },
                "matic_price": {
                    "type": "string"
                },
                "mmt_price": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
    }
}